Move BG functions to new submodule
